Wire the meta/ public-API subsystem and mtk stragglers into the build
The mutter_port/meta/ subsystem (mutter's public API header layer: window,
display, compositor, monitor, workspace, selection, etc.) existed on disk
with its own mod.rs but was never declared in mutter_port/mod.rs, and three
complete mtk ports (anonymous_file, dispose_bin, macros) were likewise
undeclared — so none of them were compiled or type-checked.

Declare `pub mod meta;` and wire the three mtk modules in, so the whole
mutter_port tree is now build-validated. Fix the errors this surfaced:

- mtk: add an `MtkRectangle` alias for `mtk::rectangle::Rectangle` (the
  meta/ files import the upstream `MtkRectangle` spelling).
- mtk/anonymous_file: compare the mmap result against `MAP_FAILED as i64`
  (the syscall returns i64; MAP_FAILED is isize).
- mtk/dispose_bin: drop the `Debug` derive from `DisposeBin`, whose
  `Box<dyn FnOnce()>` entries are not `Debug`.

Full kernel `cargo check` is green with the entire mutter_port tree wired in.

+++ src/mutter_port/mtk/anonymous_file.rs
@@ -148,7 +148,7 @@ fn sys_mmap(addr: *mut u8, length: usize, prot: i32, flags: i32, fd: i32, offset
     {
         // SAFETY: addr may be NULL; fd must be valid; offset page-aligned.
         let raw = unsafe { syscall6(SYS_MMAP, addr as i64, length as i64, prot as i64, flags as i64, fd as i64, offset) };
-        if raw == MAP_FAILED { Err(AnonymousFileError::MmapFailed(0)) } else { Ok(raw as *mut u8) }
+        if raw == MAP_FAILED as i64 { Err(AnonymousFileError::MmapFailed(0)) } else { Ok(raw as *mut u8) }
     }
     #[cfg(not(target_arch = "x86_64"))]
     {
